requirements.txt
# Astro Runtime-compatible Airflow providers
apache-airflow-providers-databricks==3.4.0
apache-airflow-providers-snowflake

# Astronomer Cosmos for dbt integration
astronomer-cosmos==1.1.3

# dbt core + Snowflake adapter (must match Cosmos compatibility matrix)
dbt-core==1.6.6
dbt-snowflake==1.6.6

apache-airflow-providers-docker